refactor: fix viz theme not initialised in storybook
